[DDPB-4477] add button to process Lay and Org CSVs in Digideps via a background task (#1001)
* [DDPB-4477] add button to process Lay and Org CSVs in Digideps via a
background task

Added a fake csv file to the client side so we can import this into
localstack S3 bucket. We also run a command in the background to process
the CSVs. This background process is linked to the
KernelEvents::TERMINATE listener. So the task will keep running until
either it finishes or the server shuts down.

Keep in mind deploying to prod while this task is running will most
likely mean we have to run it again.