Migration to align entities with database
This includes any harmless changes Doctrine wants to make, as
well as any legitimate changes where the database structure
has drifted from the field definitions in the entities. In
cases where there are cascade deletes in the db but not in
the entities, the entity relationships have been changed through
ORM annotations so that the relationships in the db remain.