Allow dual deputies for a client
Create a new client only if a deputy in a CSV row should not
be able to see the report for an existing client.

If there is an existing client with the same report type (e.g. 102)
as the deputy being added in the CSV (e.g. OPG102/pfa), then make
the deputy in the CSV a co-deputy on the existing report.

If not, create a new client and report associated with that
deputy in the CSV.