Add extra explanation about how we find multi-client candidates
I considered adding a clause to only include existing clients
in the list of candidates for multi-client. However, this discounts
multi-clients occurring in the CSV file which we have never seen
before and are not represented in the database, e.g. the same new
deputy assigned to two different clients in a single CSV upload.