Remove persister
For memory reasons, we can't create the entities and then persist
them later: for example, we have to create the court order then
directly insert into tables using its ID, rather than retrieving the
Doctrine Entity and using that. Due to the number of reports and
deputies which might be on a court order, the memory requirements are
too high.

For that reason, responsibility for creating and persisting entities
will move to the candidate converter (within the builder). This
is messier, but hopefully will allow us to process the required
number of court order changes.