Don't include the order insert action in the iterator
As we need to either insert or find the order, this is made
more complicated if we include that candidate in the iterator.

If we move it out, we can get that first and action it, then
other candidates can make use of the order ID (either from
the newly-created order, or by lookup).