DDLS-552 Deputyships builder implementation (#1972)
* Sort candidates before returning them to the builder

* Yield iterables from the builder and persister

* Loop over candidates, group them, plus skeleton to convert them to entities

* Refactor candidate action to enum

* Fix test deprecation

* Don't convert candidates if there are none to convert

* Remove unused enum

* First pass at deputyships builder

* Associate court order with at most one NDR

* Associate court orders with candidate NDRs

* Make all non-essential fields nullable

* Remove redundant factory

* Add first few tests for DeputyshipCandidateConverter

* Unit tests for updating deputy <-> order status

* Unit tests for associating deputy with order

* Refactor to a result object being returned by builder

* Fix phpstan issues

* Enforce action and courtOrderUid when creating candidate objects

* Trim trailing whitespace on save

* Complete unit tests for DeputyshipCandidateConverter

* Fix phpstan error

* Update method signatures to use Traversable

* Include court order UID in builder result

* Include outcome in builder result

* Log builder results

* Fix bug in deputyship builder

* Add unit test for deputyship builder

* Read in all candidates (will need to be a generator later)

* Extra lazily load associations with other objects

* Use array result for memory efficiency when building from candidates

* Strip out over-complicated builder implementation

* Remove persister

For memory reasons, we can't create the entities and then persist
them later: for example, we have to create the court order then
directly insert into tables using its ID, rather than retrieving the
Doctrine Entity and using that. Due to the number of reports and
deputies which might be on a court order, the memory requirements are
too high.

For that reason, responsibility for creating and persisting entities
will move to the candidate converter (within the builder). This
is messier, but hopefully will allow us to process the required
number of court order changes.

* Fix a couple of things missed during refactor

* DDLS-619 fix api side vulnerabilities (#1970)

* Fix phpstan issues

* Encapsulate ordering of candidates when processed by builder

Because we have to do the order insert first, then insert other
relationship entities, then update statuses, candidates (once
grouped by court order UID) have to be processed in a specific
order. Encapsulate ordering of candidates required by the
builder.

* Rename class for consistency

* Improve logging during build step

* Hook up command output to logger during ingest

* Log number of records loaded from CSV

* Improve logging layout

* Fix tests for changes to logging

* Return null instead of throwing an exception

* Add tests for candidate group

* Remove useless test and standardise use of mock counter

* Don't include the order insert action in the iterator

As we need to either insert or find the order, this is made
more complicated if we include that candidate in the iterator.

If we move it out, we can get that first and action it, then
other candidates can make use of the order ID (either from
the newly-created order, or by lookup).

* Stub out converter methods

* When doing DISTINCT, don't include record ID

We need to de-duplicate candidates, but distinct will treat two
otherwise identical lines as non-identical if their id field
differs.

Exclude id from the select so we can properly de-duplicate
candidates before building them.

---------

Co-authored-by: Jim <58939809+jamesrwarren@users.noreply.github.com>