DDLS-860 Cascade deputy delete to court_order_deputy (#2059)
* Add integration test that deleting a deputy deletes records in court_order_deputy

* ApiBaseTestCase doesn't need to extend WebTestCase

None of the methods of WebTestCase are used in the subclasses
of this base class, so just inherit from the KernelTestCase.

This makes the base class much lighter and more useful, potentially
replacing setup and teardown boilerplate in other tests...

* ...like this one

Where we can now get rid of boilerplate because we're using
a useful base test superclass.

* Reproduce bug in integration test

* Cascade delete court_order_deputy when deputy is removed

* Reinstate afterClass teardown

* Clean up unused member var