Use getLastLoggedIn() method rather than lastLoggedIn property
This is because we are using the codeputy template in two contexts:

* One where we are treating a user object as a proxy for a deputy,
where the lastLoggedIn property is available directly.
* One where we are actually using a court order's deputy object,
which doesn't have a lastLoggedIn property.

By replacing the property with a method call, we can proxy the
lastLoggedIn datetime of the user through its associated deputy
object. This lets us use the codeputy template with a real deputy
or with a user.