Move all integration tests to shared base class
This allows us to remove a lot of boilerplate, standardise
the style of the tests, and be more consistent in how and when
the database is purged.

A central change is to purge all db tables in the class teardown
of all tests, only leaving the dd_user table untouched. This allows
tests to be run in any order, and for the whole integration test
folder to be run at once. It also allows any integration test
to be run independently of any other.