DDLS-568 API integration tests should not be order-dependent (#2061)
* Suites

* Fix failing test

* Fix failing test in QueryTest

* Do not insert client here

* Add client to db during test

* Use ApiBaseTestCase to remove some boilerplate from test

* Remove unnecessary change to Makefile

* Call ORMPurger after every controller class test has executed

* Add purger to AbstractTestController

* Cannot purge after every controller test

While this works when running tests individually, the fixtures
are only loaded once for the entire run of a suite, so clearing
them after one test class has been run breaks subsequent tests
which rely on the fixtures.

* Return API integration test runner to former configuration

* Move all integration tests to shared base class

This allows us to remove a lot of boilerplate, standardise
the style of the tests, and be more consistent in how and when
the database is purged.

A central change is to purge all db tables in the class teardown
of all tests, only leaving the dd_user table untouched. This allows
tests to be run in any order, and for the whole integration test
folder to be run at once. It also allows any integration test
to be run independently of any other.

* Prevent warning about missing array key

* Remove reference to non-existent resultset field

---------

Co-authored-by: Elliot Smith <elliot.smith2@justice.gov.uk>