DDLS-852 Court order co-deputy display (#2065)
* Fix e2e test definitions and calls

* Skeleton for first end-to-end test

* Flesh out test for court order co-deputy display

* fix-test

* Deputy must be associated with user for court orders to work

* Include logged in deputy in list of co-deputies for court order

This is to ensure parity with current client display, which shows
the logged-in deputy as well as other co-deputies.

* Show co-deputies for court order based on number of deputies on order, not client

* Include user with court order deputy

* Use getLastLoggedIn() method rather than lastLoggedIn property

This is because we are using the codeputy template in two contexts:

* One where we are treating a user object as a proxy for a deputy,
where the lastLoggedIn property is available directly.
* One where we are actually using a court order's deputy object,
which doesn't have a lastLoggedIn property.

By replacing the property with a method call, we can proxy the
lastLoggedIn datetime of the user through its associated deputy
object. This lets us use the codeputy template with a real deputy
or with a user.

* Proxy user->lastLoggedIn through deputy method

* Ensure that NDR template displays correctly for court order

* Check that co-deputy is displayed correctly

* Use more correct variable names

* Add test for display of registered court order co-deputies