DDLS-907 Add missing deputy_case entries (#2111)
* Make orderUid field nullable and add userId field

This is to support adding deputy_case candidates during
deputyships ingest.

* Enable creation of deputy_case candidates

* Add court order UID to all candidates

Court order UID is used for sorting candidates in the builder,
so removing it might cause various problems. For the sake
of simplicity, when creating a deputy_case candidate, associate
it with a court order UID so it is processed with the other
candidates for that court order.

* Expose factory so it can be integration tested

* Add integration test for deputy_case candidates factory

* Add implementation for deputy_case candidates factory

* Add deputy_case associations during lay ingest

* Make query more efficient

* Use pre-reg table to find missing deputy+case associations

* Save missing deputy_case records to db

* Add missing deputy_case associations during lay ingest

* Remove unnecessary and unused classes

* Decouple event subscribers and handle errors gracefully

The event subscriber which logs audit messages via Monolog was
throwing an uncaught AWS exception; this then prevented the
user's active and registration_date fields from being updated.

* Decouple the audit log from the subscriber which updates the
registered user's fields.
* Capture any errors thrown while sending messages to the audit
log, so that subsequent subscribers are not interrupted/broken.

* Remove redundant service definition

* Fix phpstan API errors

* Add new service to command test

---------

Co-authored-by: Jack Goodby <jackgoodby@gmail.com>
Co-authored-by: Richard McHale <legionsofbob@gmail.com>