DDLS-961 Fix incorrect data persist, flush, and clear sequencing (#2134)
* Only call clear() after whole collection of entities is done

* Handle all exceptions during lay CSV ingest

* Use early return instead of massive conditional body

* Flush and clear after each client

* Integration test for LayRegistrationService

* Produce coverage report files when running single integration test

* Fix docs

* Add function to create PreRegistration fixtures in integration tests

* Add test for adding a missing hybrid report to a client

* Add integration test for database persists and flushes

* Client must be persisted with batches

* Fix incorrect unit test

* Enable deputy UID to be set on prereg fixture and don't persist in Fixtures class

Other Fixtures methods don't persist the created objects, so follow that
pattern here.

* Add test to check entity manager persist and flush ordering

* Reorganise persist and flush to prevent error

This error happens in the e2e tests:

A new entity was found through the relationship
'App\\Entity\\Client#deputy' that was not configured to cascade
persist operations for entity

This is the same error that occurred when adding missing reports,
and is down to objects not being persisted/flushed in the correct
order.

* Include stacktrace in exception logging