Loosen criteria for matching reports to court orders
We were not matching non-hybrid reports to hybrid deputyships.
This is incorrect because a court order may start out with a
non-hybrid report, then be changed so that it has hybrid reporting.

Remove the criteria from the report matching query so that
non-hybrid reports are attached to hybrid deputyships, and
hybrid reports are matched to non-hybrid deputyships.