DDLS-814 Streamline file upload workflow (#2144)
* Call build-js target when doing up-app-rebuild

* Reset cache when doing up-app-rebuild

* Call MAKE to bring the app down and build the JS in the correct order

* Enable all local nginx containers to handle 15M files

* Rewrite UploadFile functionality

* Don't enforce nginx max upload file size in local dev env

This can interfere with the JS we are using to show nicer
error messages if a file is too large.

* If file is too big, return false so form does not submit

* Show messages as requested in ticket

* Disable file select button after form submit

* Add attributes so UploadFile works correctly

* Add attributes so UploadFile works correctly

* Lint fixes for JS

* Refactor UploadFile so it can be tested more easily

* Make function unit testable

* Add unit test for UploadFile.js

* Use JS to hide the file chooser submit

If JS is broken or otherwise unavailable, the submit button
should remain visible.

* Remove extra 'Save' button on upload form

* Enable file upload JS later in bootstrap

Also fix casing of imported module.

* Remove unnecessary check for directory

This check prevents the npm install, lint, audit, and
webpack compilation if the directory has already been
built, presumably during a previous action run.

Remove this so that we always carry out these steps.

If the webpack build, npm install etc. are cached, they can decide
for themselves whether to run again.