Remove more irrelevant tests
These tests were to check the situation where a user logged in
and chose a client, and was then presented with pages for that
client. To change client, they would have to return to the choose
a client page.

As we have removed this functionality, because you are now
looking at court orders and not choosing a client context, these
tests are irrelevant.