Reverse the test for checking update deputy status on order
We should check making a deputy inactive on an order, because
this is where the bug occurs.

Using true as a parameter results in a valid SQL update query, while
using false results in an invalid query. This causes problems
where we are setting the deputy status on an order to false.