DDLS-1036 Fix deputy status update during deputyship ingest (#2184)
* Add unit test for creating an update deputy status on order candidate

* Model integration tests were not being run

* Reverse the test for checking update deputy status on order

We should check making a deputy inactive on an order, because
this is where the bug occurs.

Using true as a parameter results in a valid SQL update query, while
using false results in an invalid query. This causes problems
where we are setting the deputy status on an order to false.

* Fix parameter passed to update query

* Remove override of tearDownAfterClass

* Remove unused import

* Remove unnecessary variable

* Slightly clean up other insert query with a boolean to make it consistent