Allow generated deputy to share the name of the user
In some cases, tests will compare details of the logged in
user with page text. If we are really looking at deputy details,
we need to make sure that the user and the deputy line up.
By adding this, we can ask that a generated deputy's details
line up with the user's.