Lay home page is more variable now
Whereas before the lay home page would be a client/ID page,
it may now be courtorder/UID, courtorder/waiting, or
courtorder/choose-a-court-order.

Modify the test helper so that any of these pages is acceptable
(as any of them could be lay home, depending on the deputy and
the state of their data).