DDLS-808/DDLS-1087 Route users to court order pages (revised) (#2219)
* Simplification and cleanup of redirector

* Redirect to court order page after login, once details are complete

* Redirect from choose a court order to single court order

* Use redirects for both the single and waiting court order pages

This makes it easier to see what's going wrong in the tests.

It should also make it easier to track down problems reported
by users (as it indicates where they have no court orders in the
URL).

* Move secondary user redirection to event subscriber

* Only redirect non-primary logged in users if they are a lay

As this check was previously on the choose a client and client
pages, assume it only applies to lay deputies.

* Only protect /courtorder/* pages against secondary users

* Fix config for LoginSubscriber

* Clean up code for LoginSubscriber

* Fix unit tests

* Fix e2e test for court order with changed report type

* Ensure new user fixture gets a deputy record

* Clean up how we fetch user and deputy fixtures

* Need to find a different way to create missing deputies

* Fix another e2e court order test

* Fix remaining court order e2e tests

* Cleanup style

* Add missing trait to context

* Fix NDR login e2e test

* Remove tests for discharging deputies

After discussion with the product team, we agreed that these
tests are unnecessary as we eventually plan to remove the
discharge functionality from digideps. This will be handled
during the deputyship ingest instead.

* Remove e2e test covered by court order invite tests

* Fix report management e2e tests

* Redundant test, as already tested by court order page e2e tests

* Fix more e2e tests

* Remove redundant e2e test (covered by court order tests)

* Change tag to make running this one feature easier

* Remove unused member variables and pointless private method

* Remove unused code

* Get fixtures to properly associate reports and ndrs with court order

* Fix e2e test

* Fix court order e2e tests again

* Fix remaining e2e tests

* Fix co-deputy e2e test to use court order invite

* Fix imports for e2e tests

* Fix smoke tests

* Reset client phpstan baseline

* Reset API phpstan baseline

* Add missing function to CourtOrderTrait

* Remove irrelevant test (NDRs not visible with court orders)

* Remove another e2e test about NDRs with court orders (there won't be any)

* Fix broken fixture method

* Remove unused argument from function call for fixture

* Allow generated deputy to share the name of the user

In some cases, tests will compare details of the logged in
user with page text. If we are really looking at deputy details,
we need to make sure that the user and the deputy line up.
By adding this, we can ask that a generated deputy's details
line up with the user's.

* Fix broken e2e test

* Fix routing for user research controller

* Fix redirects for ReportController

* Fix routing to lay_home

* Lay home page is more variable now

Whereas before the lay home page would be a client/ID page,
it may now be courtorder/UID, courtorder/waiting, or
courtorder/choose-a-court-order.

Modify the test helper so that any of these pages is acceptable
(as any of them could be lay home, depending on the deputy and
the state of their data).

* Remove invalid test assertion

User will not see their reports after registering until
the ingest has associated them with their court orders.

* Remove redundant test

We no longer prevent multiple invites being sent to a co-deputy,
so this error message is never shown, and this validation never
happens for invitations to court orders.

* Fix phpstan issue

* Reorder Redirect methods to enable diff to be reviewed more easily

* Fix typo

---------

Co-authored-by: Jim <58939809+jamesrwarren@users.noreply.github.com>
Co-authored-by: Raffers <Chris.rafferty@digital.justice.gov.uk>